requirements.txt
customtkinter
requests
Pillow
pystray